+++ R/handler.R
@@ -7,7 +7,7 @@
 #' @param precision If `trim` is `TRUE`, the total number of significant digits to keep
 #'   for each result or the number of digits after the decimal place otherwise.
 #' @param trim Use `FALSE` to keep trailing zeroes after the decimal place.
-#' @param x A vector that can be interpreted as a [wkb()], [wkt()], or [wksxp()].
+#' @param x A geometry vector (e.g., [wkb()] or [wkt()]).
 #' @param ... Unused
 #' @return A WK handler.
 #' @export