+++ R/translate.R
@@ -4,7 +4,6 @@
 #' @param wkb A `list()` of [raw()] vectors, such as that
 #'   returned by `sf::st_as_binary()`.
 #' @param wkt A character vector containing well-known text.
-#' @param wksxp A `list()` of classed objects
 #' @param trim Trim unnecessary zeroes in the output?
 #' @param precision The rounding precision to use when writing
 #'   (number of decimal places).
@@ -22,10 +21,8 @@
 #'
 #' @return `*_translate_wkt()` returns a character vector of
 #'   well-known text; `*_translate_wkb()` returns a list
-#'   of raw vectors, and `*_translate_wksxp()` returns an unclassed
-#'   list of [wksxp()] geometries. Unlike [as_wkb()], [as_wkt()], and
-#'   [as_wksxp()], these functions do not attach
-#'   a class to the output.
+#'   of raw vectors. Unlike [as_wkb()], [as_wkt()],
+#'   these functions do not attach a class to the output.
 #'
 #' @export
 #'
@@ -66,17 +63,6 @@ wkb_translate_wkb <- function(wkb, include_z = NA, include_m = NA, include_srid 
 
 #' @rdname wkb_translate_wkt
 #' @export
-wkb_translate_wksxp <- function(wkb, include_z = NA, include_m = NA, include_srid = NA) {
-  cpp_wkb_translate_wksxp(
-    wkb,
-    includeZ = include_z,
-    includeM = include_m,
-    includeSRID = include_srid
-  )
-}
-
-#' @rdname wkb_translate_wkt
-#' @export
 wkt_translate_wkt <- function(wkt, include_z = NA, include_m = NA, include_srid = NA,
                               precision = 16, trim = TRUE) {
   cpp_wkt_translate_wkt(
@@ -103,56 +89,6 @@ wkt_translate_wkb <- function(wkt, include_z = NA, include_m = NA, include_srid 
   )
 }
 
-#' @rdname wkb_translate_wkt
-#' @export
-wkt_translate_wksxp <- function(wkt, include_z = NA, include_m = NA, include_srid = NA) {
-  cpp_wkt_translate_wksxp(
-    wkt,
-    includeZ = include_z,
-    includeM = include_m,
-    includeSRID = include_srid
-  )
-}
-
-#' @rdname wkb_translate_wkt
-#' @export
-wksxp_translate_wkt <- function(wksxp, include_z = NA, include_m = NA, include_srid = NA,
-                                 precision = 16, trim = TRUE) {
-  cpp_wksxp_translate_wkt(
-    wksxp,
-    includeZ = include_z,
-    includeM = include_m,
-    includeSRID = include_srid,
-    precision = precision,
-    trim = trim
-  )
-}
-
-#' @rdname wkb_translate_wkt
-#' @export
-wksxp_translate_wkb <- function(wksxp, include_z = NA, include_m = NA, include_srid = NA,
-                                endian = wk_platform_endian(), buffer_size = 2048) {
-  cpp_wksxp_translate_wkb(
-    wksxp,
-    includeZ = include_z,
-    includeM = include_m,
-    includeSRID = include_srid,
-    endian = endian,
-    bufferSize = buffer_size
-  )
-}
-
-#' @rdname wkb_translate_wkt
-#' @export
-wksxp_translate_wksxp <- function(wksxp, include_z = NA, include_m = NA, include_srid = NA) {
-  cpp_wksxp_translate_wksxp(
-    wksxp,
-    includeZ = include_z,
-    includeM = include_m,
-    includeSRID = include_srid
-  )
-}
-
 # ----- leaving these unexported until the arguments are stable
 
 xyzm_translate_wkt <- function(xyzm, precision = 16, trim = TRUE) {
@@ -163,10 +99,6 @@ xyzm_translate_wkb <- function(xyzm, endian = wk_platform_endian(), buffer_size 
   cpp_translate_xyzm_wkb(xyzm, endian, bufferSize = buffer_size)
 }
 
-xyzm_translate_wksxp <- function(xyzm) {
-  cpp_translate_xyzm_wksxp(xyzm)
-}
-
 rct_translate_wkt <- function(rct, precision = 16, trim = TRUE) {
   cpp_translate_rct_wkt(rct, precision, trim)
 }
@@ -175,20 +107,12 @@ rct_translate_wkb <- function(rct, endian = wk_platform_endian(), buffer_size = 
   cpp_translate_rct_wkb(rct, endian, bufferSize = buffer_size)
 }
 
-rct_translate_wksxp <- function(rct) {
-  cpp_translate_rct_wksxp(rct)
-}
-
 wkt_translate_xyzm <- function(wkt, include_z = NA, include_m = NA) {
   xyzm_trim(cpp_translate_wkt_xyzm(wkt), include_z, include_m)
 }
 
 wkb_translate_xyzm <- function(wkb, include_z = NA, include_m = NA) {
   xyzm_trim(cpp_translate_wkb_xyzm(wkb), include_z, include_m)
-}
-
-wksxp_translate_xyzm <- function(wksxp, include_z = NA, include_m = NA) {
-  xyzm_trim(cpp_translate_wksxp_xyzm(wksxp), include_z, include_m)
 }
 
 xyzm_trim <- function(result, include_z, include_m) {