+++ data-raw/gen-vctrs-compat.R
@@ -20,12 +20,6 @@ vec_ptype2.{ CLASS }.wk_wkt <- function(x, y, ..., x_arg = "x", y_arg = "y") {{
   new_wk_PICK(crs = wk_crs_output(x, y))
 }}
 
-#\' @method vec_ptype2.{ CLASS } wk_wksxp
-#\' @export
-vec_ptype2.{ CLASS }.wk_wksxp <- function(x, y, ..., x_arg = "x", y_arg = "y") {{
-  new_wk_PICK(crs = wk_crs_output(x, y))
-}}
-
 #\' @method vec_ptype2.{ CLASS } wk_xy
 #\' @export
 vec_ptype2.{ CLASS }.wk_xy <- function(x, y, ..., x_arg = "x", y_arg = "y") {{
@@ -68,12 +62,6 @@ vec_cast.{ CLASS }.wk_wkb <- function(x, to, ...) {{
 #\' @method vec_cast.{ CLASS } wk_wkt
 #\' @export
 vec_cast.{ CLASS }.wk_wkt <- function(x, to, ...) {{
-  as_{ gsub("wk_", "", CLASS) }(x, crs = wk_crs_output(x, to))
-}}
-
-#\' @method vec_cast.{ CLASS } wk_wksxp
-#\' @export
-vec_cast.{ CLASS }.wk_wksxp <- function(x, to, ...) {{
   as_{ gsub("wk_", "", CLASS) }(x, crs = wk_crs_output(x, to))
 }}