+++ src/cpp11.cpp
@@ -24,28 +24,18 @@ extern "C" SEXP _wk_wk_cpp_wkt_writer(SEXP precision, SEXP trim) {
 extern "C" {
 /* .Call calls */
 extern SEXP _wk_cpp_format_wkb(SEXP, SEXP, SEXP, SEXP);
-extern SEXP _wk_cpp_format_wksxp(SEXP, SEXP, SEXP, SEXP);
 extern SEXP _wk_cpp_format_wkt(SEXP, SEXP, SEXP, SEXP);
 extern SEXP _wk_cpp_problems_wkb(SEXP);
-extern SEXP _wk_cpp_problems_wksxp(SEXP);
 extern SEXP _wk_cpp_problems_wkt(SEXP);
 extern SEXP _wk_cpp_translate_rct_wkb(SEXP, SEXP, SEXP);
-extern SEXP _wk_cpp_translate_rct_wksxp(SEXP);
 extern SEXP _wk_cpp_translate_rct_wkt(SEXP, SEXP, SEXP);
 extern SEXP _wk_cpp_translate_wkb_xyzm(SEXP);
-extern SEXP _wk_cpp_translate_wksxp_xyzm(SEXP);
 extern SEXP _wk_cpp_translate_wkt_xyzm(SEXP);
 extern SEXP _wk_cpp_translate_xyzm_wkb(SEXP, SEXP, SEXP);
-extern SEXP _wk_cpp_translate_xyzm_wksxp(SEXP);
 extern SEXP _wk_cpp_translate_xyzm_wkt(SEXP, SEXP, SEXP);
 extern SEXP _wk_cpp_wkb_translate_wkb(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
-extern SEXP _wk_cpp_wkb_translate_wksxp(SEXP, SEXP, SEXP, SEXP);
 extern SEXP _wk_cpp_wkb_translate_wkt(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
-extern SEXP _wk_cpp_wksxp_translate_wkb(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
-extern SEXP _wk_cpp_wksxp_translate_wksxp(SEXP, SEXP, SEXP, SEXP);
-extern SEXP _wk_cpp_wksxp_translate_wkt(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _wk_cpp_wkt_translate_wkb(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
-extern SEXP _wk_cpp_wkt_translate_wksxp(SEXP, SEXP, SEXP, SEXP);
 extern SEXP _wk_cpp_wkt_translate_wkt(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _wk_wk_cpp_handle_wkt(SEXP, SEXP);
 extern SEXP _wk_wk_cpp_wkt_writer(SEXP, SEXP);
@@ -60,28 +50,18 @@ extern SEXP wk_c_wkb_writer_new();
 
 static const R_CallMethodDef CallEntries[] = {
     {"_wk_cpp_format_wkb",            (DL_FUNC) &_wk_cpp_format_wkb,            4},
-    {"_wk_cpp_format_wksxp",          (DL_FUNC) &_wk_cpp_format_wksxp,          4},
     {"_wk_cpp_format_wkt",            (DL_FUNC) &_wk_cpp_format_wkt,            4},
     {"_wk_cpp_problems_wkb",          (DL_FUNC) &_wk_cpp_problems_wkb,          1},
-    {"_wk_cpp_problems_wksxp",        (DL_FUNC) &_wk_cpp_problems_wksxp,        1},
     {"_wk_cpp_problems_wkt",          (DL_FUNC) &_wk_cpp_problems_wkt,          1},
     {"_wk_cpp_translate_rct_wkb",     (DL_FUNC) &_wk_cpp_translate_rct_wkb,     3},
-    {"_wk_cpp_translate_rct_wksxp",   (DL_FUNC) &_wk_cpp_translate_rct_wksxp,   1},
     {"_wk_cpp_translate_rct_wkt",     (DL_FUNC) &_wk_cpp_translate_rct_wkt,     3},
     {"_wk_cpp_translate_wkb_xyzm",    (DL_FUNC) &_wk_cpp_translate_wkb_xyzm,    1},
-    {"_wk_cpp_translate_wksxp_xyzm",  (DL_FUNC) &_wk_cpp_translate_wksxp_xyzm,  1},
     {"_wk_cpp_translate_wkt_xyzm",    (DL_FUNC) &_wk_cpp_translate_wkt_xyzm,    1},
     {"_wk_cpp_translate_xyzm_wkb",    (DL_FUNC) &_wk_cpp_translate_xyzm_wkb,    3},
-    {"_wk_cpp_translate_xyzm_wksxp",  (DL_FUNC) &_wk_cpp_translate_xyzm_wksxp,  1},
     {"_wk_cpp_translate_xyzm_wkt",    (DL_FUNC) &_wk_cpp_translate_xyzm_wkt,    3},
     {"_wk_cpp_wkb_translate_wkb",     (DL_FUNC) &_wk_cpp_wkb_translate_wkb,     6},
-    {"_wk_cpp_wkb_translate_wksxp",   (DL_FUNC) &_wk_cpp_wkb_translate_wksxp,   4},
     {"_wk_cpp_wkb_translate_wkt",     (DL_FUNC) &_wk_cpp_wkb_translate_wkt,     6},
-    {"_wk_cpp_wksxp_translate_wkb",   (DL_FUNC) &_wk_cpp_wksxp_translate_wkb,   6},
-    {"_wk_cpp_wksxp_translate_wksxp", (DL_FUNC) &_wk_cpp_wksxp_translate_wksxp, 4},
-    {"_wk_cpp_wksxp_translate_wkt",   (DL_FUNC) &_wk_cpp_wksxp_translate_wkt,   6},
     {"_wk_cpp_wkt_translate_wkb",     (DL_FUNC) &_wk_cpp_wkt_translate_wkb,     6},
-    {"_wk_cpp_wkt_translate_wksxp",   (DL_FUNC) &_wk_cpp_wkt_translate_wksxp,   4},
     {"_wk_cpp_wkt_translate_wkt",     (DL_FUNC) &_wk_cpp_wkt_translate_wkt,     6},
     {"_wk_wk_cpp_handle_wkt",         (DL_FUNC) &_wk_wk_cpp_handle_wkt,         2},
     {"_wk_wk_cpp_wkt_writer",         (DL_FUNC) &_wk_wk_cpp_wkt_writer,         2},